+++ page-categories.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Template Page of Category
  *
@@ -6,7 +7,7 @@
  */
 ?>
 <?php if (!defined('__TYPECHO_ROOT_DIR__')) exit; ?>
-<?php $this->need('header.php'); ?>
+<?php $this->need('component/header.php'); ?>
 
 <div id="main" class="main" role="main">
     <div class="main-inner clearfix">
@@ -46,47 +47,9 @@
         border: none
     }
 
-    .category-post-wrap {
-        display: inline-block;
-        width: 33.33%;
-        box-sizing: border-box;
-        float: left;
-        padding: 10px;
-    }
-
-    @media (max-width: 991px) {
-        .category-post-wrap {
-            width: 50%;
-        }
-    }
-
-    .category-post {
-        background-color: #fff;
-        padding: 20px;
-        box-shadow: 0 1px 4px rgba(0, 0, 0, .04);
-    }
-
     .content-wrap {
         float: none
     }
-
-    .category-post-title {
-        font-weight: normal;
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-        font-size: 14px;
-        margin: 0;
-    }
-
-    .category-title {
-        margin: 10px;
-        font-size: 14px;
-    }
-
-    .category-post-item {
-        animation: fade-in-top .3s .3s backwards;
-    }
 </style>
 
-<?php $this->need('footer.php'); ?>
+<?php $this->need('component/footer.php'); ?>