+++ page-link.php
@@ -8,11 +8,11 @@
 ?>
 <?php if (!defined('__TYPECHO_ROOT_DIR__')) exit; ?>
 
-<?php $this->need('header.php'); ?>
+<?php $this->need('component/header.php'); ?>
 
 <div id="main" class="main" role="main">
   <div class="main-inner clearfix">
-    <?php if (isPc()) $this->need('sidebar.php'); ?>
+    <?php if (isPc()) $this->need('component/sidebar.php'); ?>
     <div class="content-wrap">
       <article class="post" itemscope itemtype="http://schema.org/BlogPosting">
         <h1 class="post-title" itemprop="name headline"><a class="post-title-link" itemprop="url" href="<?php $this->permalink() ?>"><?php $this->title() ?></a></h1>
@@ -20,25 +20,10 @@
           <?php $this->content(); ?>
         </div>
       </article>
-      <?php $this->need('comments.php'); ?>
+      <?php $this->need('component/comments.php'); ?>
     </div>
   </div>
 </div>
 
 
-<?php $this->need('footer.php'); ?>
-
-<script>
-  var warp = document.getElementById('link'),
-    br = [].slice.call(warp.getElementsByTagName('BR'))
-    itemLink = [].slice.call(document.querySelectorAll('#link ul li'))
-
-  br.forEach((v, k) => {
-    v.parentElement.removeChild(v)
-  })
-  itemLink.forEach((v, k) => {
-    if (!!!v.querySelector('img')) {
-      v.style.paddingTop = '10px'
-    }
-  })
-</script>
+<?php $this->need('component/footer.php'); ?>