+++ page.php
@@ -1,9 +1,9 @@
 <?php if (!defined('__TYPECHO_ROOT_DIR__')) exit; ?>
-<?php $this->need('header.php'); ?>
+<?php $this->need('component/header.php'); ?>
 
 <div id="main" class="main" role="main">
     <div class="main-inner clearfix">
-        <?php if (isPc()) $this->need('sidebar.php'); ?>
+        <?php if (isPc()) $this->need('component/sidebar.php'); ?>
         <div class="content-wrap">
             <article class="post" itemscope itemtype="http://schema.org/BlogPosting">
                 <h1 class="post-title" itemprop="name headline"><a class="post-title-link" itemprop="url" href="<?php $this->permalink() ?>"><?php $this->title() ?></a></h1>
@@ -11,10 +11,10 @@
                     <?php $this->content(); ?>
                 </div>
             </article>
-            <?php $this->need('comments.php'); ?>
+            <?php $this->need('component/comments.php'); ?>
         </div>
     </div>
 </div>
 
 
-<?php $this->need('footer.php'); ?>
+<?php $this->need('component/footer.php'); ?>