+++ post.php
@@ -1,59 +1,39 @@
 <?php if (!defined('__TYPECHO_ROOT_DIR__')) exit; ?>
-<?php $this->need('header.php'); ?>
+<?php $this->need('component/header.php'); ?>
 
 <div id="main" role="main">
     <div class="main-inner clearfix">
         <div class="content-wrap">
-            <article class="post" itemscope itemtype="http://schema.org/BlogPosting">
-                <div class="post-badge">
-                    <span itemprop="about" itemscope="" itemtype="http://schema.org/Thing">
-                        <?php if($this->category) $this->category(','); else _e('o_o ....')  ?>
-                    </span>
-                </div>
-                <header class="post-header">
-                    <h1 class="post-title" itemprop="name headline"><?php $this->title() ?></h1>
-                    <div class="post-meta">
-                        <span><?php _e('<i class="iconfont icon-time"></i> 发表于 '); ?><time datetime="<?php $this->date('c'); ?>" itemprop="datePublished"><?php $this->date(); ?></time></span>
-                        <span><?php _e('<i class="iconfont icon-eye"></i> 浏览量 '); getPostView($this); ?></span>
-                        <span itemprop="interactionCount"><a itemprop="discussionUrl" href="<?php $this->permalink() ?>#comments"><?php $this->commentsNum('<i class="iconfont icon-Comments"></i> 没有评论', '<i class="iconfont icon-Comments"></i> 评论数 1', '<i class="iconfont icon-Comments"></i> 评论数 %d'); ?></a></span>
-                        <span><?php if($this -> authorId == $this -> user -> uid): ?>
-                            <a class="edit-link" href="<?php $this -> options -> adminUrl(); ?>write-post.php?cid=<?php echo $this -> cid; ?>" target="_blank">编辑</a>
-                            <?php endif; ?>
-                        </span>
+            <?php $this->need('component/post-card.php'); ?>
+            <?php if ($this->options->WechatQR || $this->options->AlipayQR) : ?>
+                <div class="free-reward">
+                    <button class="btn btn-pay"><?php i18n('赞赏') ?></button>
+                    <div class="qr" style="height:0">
+                        <?php if ($this->options->WechatQR) : ?>
+                            <div id="wechat">
+                                <img id="wechat-qr" src="<?php $this->options->WechatQR() ?>" alt="<?php i18n('微信支付') ?>" data-action="zoom">
+                                <p><?php i18n('微信支付') ?></p>
+                            </div>
+                        <?php endif; ?>
+                        <?php if ($this->options->AlipayQR) : ?>
+                            <div id="alipay">
+                                <img id="qq-alipay" src="<?php $this->options->AlipayQR() ?>" alt="<?php i18n('支付宝支付') ?>" data-action="zoom">
+                                <p><?php i18n('支付宝支付') ?></p>
+                            </div>
+                        <?php endif; ?>
                     </div>
-                </header>
-                <div class="post-content" itemprop="articleBody">
-                    <?php replaceTag($this->content, $this->user->hasLogin()); ?>
                 </div>
-                <?php if ($this->options->WechatQR || $this->options->AlipayQR) : ?>
-                    <div class="free-reward">
-                        <button class="btn btn-pay">赞赏</button>
-                        <div class="qr" style="height:0">
-                            <?php if ($this->options->WechatQR) : ?>
-                                <div id="wechat">
-                                    <img id="wechat-qr" src="<?php $this->options->WechatQR() ?>" alt="微信扫一扫，向我赞赏" data-action="zoom">
-                                    <p>微信扫一扫，向我赞赏</p>
-                                </div>
-                            <?php endif; ?>
-                            <?php if ($this->options->AlipayQR) : ?>
-                                <div id="alipay">
-                                    <img id="qq-alipay" src="<?php $this->options->AlipayQR() ?>" alt="微信扫一扫，向我赞赏" data-action="zoom">
-                                    <p>支付宝扫一扫，向我赞赏</p>
-                                </div>
-                            <?php endif; ?>
-                        </div>
-                    </div>
-                <?php endif; ?>
-                <div itemprop="keywords" class="tags"><?php $this->tags('', true, ''); ?></div>
-                <footer class="post-footer">
-                    <div class="post-nav-next post-nav-item"><i class="iconfont icon-prev-m"></i><?php $this->thePrev('%s', '没有了'); ?></div>
-                    <div class="post-nav-prev post-nav-item"><?php $this->theNext('%s', '没有了'); ?><i class="iconfont icon-next-m"></i></div>
-                </footer>
+            <?php endif; ?>
+            <div itemprop="keywords" class="tags"><?php $this->tags('', true, ''); ?></div>
+            <footer class="post-footer">
+                <div class="post-nav-next post-nav-item"><i class="iconfont icon-prev-m"></i><?php $this->thePrev('%s', _i18n('没有了')); ?></div>
+                <div class="post-nav-prev post-nav-item"><?php $this->theNext('%s', _i18n('没有了')); ?><i class="iconfont icon-next-m"></i></div>
+            </footer>
             </article>
-            <?php $this->need('comments.php'); ?>
+            <?php $this->need('component/comments.php'); ?>
         </div>
-        <?php if (isPc()) $this->need('sidebar.php'); ?>
+        <?php if (isPc()) $this->need('component/sidebar.php'); ?>
     </div>
 </div>
 
-<?php $this->need('footer.php'); ?>
+<?php $this->need('component/footer.php'); ?>